updated util classes and updateCutOffControler

+++ payroll/src/main/java/org/ideyatech/payroll/util/OvertimeUtil.java
@@ -1,7 +1,7 @@
 package org.ideyatech.payroll.util;
 
 public class OvertimeUtil {
-	public static double getOvertimeAmount(int overtime, double salary, int workingDays){
+	public static double getOvertimeAmount(float overtime, double salary, int workingDays){
 		return (((salary/2)/workingDays)/8)*overtime;
 	}
 }
